// src/apollo/queries/__generated__/getId.ts
/* tslint:disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: getId
// ====================================================

export interface getId_user {
  __typename: "User";
  id: string | null;
}

export interface getId {
  user: getId_user | null;
}